refactor(crypto): move encryption primitives to shared package
